Update generate_model_from_checkpoint.py (#901)
